refactor: split MP4 sample timing logic
