+++ crates/media-parser/src/format/mp4/atoms/mod.rs
@@ -22,6 +22,7 @@ mod media;
 mod moov;
 mod nav;
 mod read;
+mod sample_timing;
 mod samples;
 mod tags;
 mod types;
@@ -41,10 +42,12 @@ pub(super) use media::{
    audio_params, fourcc_string, parse_hdlr, parse_mdhd, parse_stsd, parse_tkhd, stts_sample_count,
    visual_dimensions,
 };
+pub(super) use sample_timing::{
+   CompositionOffset, PresentationTimeline, duration_to_ticks, parse_ctts, stts_duration_ticks,
+   ticks_to_duration,
+};
 pub(super) use samples::{
-   CompositionOffset, PresentationTimeline, SampleLocator, SampleSizes, StscEntry,
-   duration_to_ticks, nearest_sync_sample, next_sync_sample, parse_avc_config, parse_chunk_offsets,
-   parse_ctts, parse_sample_sizes, parse_stsc, parse_stss, range_uses_description_index,
-   sample_description_index, sample_size, stts_duration_ticks, table_entries, ticks_to_duration,
-   validate_sample_tables,
+   SampleLocator, SampleSizes, StscEntry, nearest_sync_sample, next_sync_sample, parse_avc_config,
+   parse_chunk_offsets, parse_sample_sizes, parse_stsc, parse_stss, range_uses_description_index,
+   sample_description_index, sample_size, table_entries, validate_sample_tables,
 };
