+++ crates/media-parser/src/format/mp4/atoms/samples.rs
@@ -1,11 +1,9 @@
 //! MP4 sample-table parsing and sample reads.
 
+use super::sample_timing::{CompositionOffset, stts_sample_count};
 use super::{Mp4Nav, iter_boxes};
 use crate::decoders::h264::AvcConfig;
-#[cfg(test)]
-use crate::errors::{MediaParserError, Result};
 use crate::helpers::{read_u16_be, read_u32_be, read_u64_be};
-use std::time::Duration;
 
 #[derive(Debug, Clone, Copy)]
 pub struct StscEntry {
@@ -19,142 +17,6 @@ pub struct SampleSizes {
    pub fixed_size: u32,
    pub sizes: Vec<u32>,
    pub sample_count: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct CompositionOffset {
-   pub sample_count: u32,
-   pub sample_offset: i64,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct SampleSelection {
-   pub sample_index: u32,
-   pub presentation_tick: u64,
-}
-
-#[cfg(test)]
-const MAX_SAMPLES_PER_RANGE: usize = 16_384;
-const MAX_PRESENTATION_TIMELINE_BYTES: usize = 128 * 1024 * 1024;
-
-fn presentation_timeline_sample_count_fits(sample_count: usize) -> bool {
-   let Some(bytes_per_sample) = std::mem::size_of::<i128>().checked_add(std::mem::size_of::<u32>())
-   else {
-      return false;
-   };
-   sample_count != 0
-      && sample_count
-         .checked_mul(bytes_per_sample)
-         .is_some_and(|bytes| bytes <= MAX_PRESENTATION_TIMELINE_BYTES)
-}
-
-/// Reusable presentation timestamps for a validated MP4 video sample table.
-#[derive(Debug)]
-pub struct PresentationTimeline {
-   /// Signed presentation ticks in one-based MP4 sample order.
-   ticks_by_sample: Vec<i128>,
-   /// One-based sample indices ordered by `(presentation_tick, sample_index)`.
-   samples_by_time: Vec<u32>,
-}
-
-impl PresentationTimeline {
-   pub fn new(
-      stts: &[u8],
-      composition_offsets: Option<&[CompositionOffset]>,
-      presentation_offset: i64,
-      sample_count: u32,
-   ) -> Option<Self> {
-      let sample_count = usize::try_from(sample_count).ok()?;
-      if !presentation_timeline_sample_count_fits(sample_count)
-         || composition_offsets
-            .is_some_and(|offsets| offsets.iter().any(|entry| entry.sample_count == 0))
-      {
-         return None;
-      }
-
-      let mut ticks_by_sample = Vec::new();
-      ticks_by_sample.try_reserve_exact(sample_count).ok()?;
-      let mut samples_by_time = Vec::new();
-      samples_by_time.try_reserve_exact(sample_count).ok()?;
-      let mut walker = TimingWalker::new(stts, composition_offsets)?;
-
-      loop {
-         match walker.next_segment() {
-            TimingStep::Segment(segment) => {
-               for position in 0..segment.sample_count {
-                  if ticks_by_sample.len() == sample_count {
-                     return None;
-                  }
-                  let sample_index = segment.first_sample.checked_add(position)?;
-                  let decode_tick = segment
-                     .decode_tick
-                     .checked_add(u64::from(position).checked_mul(segment.sample_delta)?)?;
-                  let presentation_tick = i128::from(decode_tick)
-                     .checked_add(i128::from(segment.composition_offset))?
-                     .checked_sub(i128::from(presentation_offset))?;
-                  ticks_by_sample.push(presentation_tick);
-                  if u64::try_from(presentation_tick).is_ok() {
-                     samples_by_time.push(sample_index);
-                  } else if presentation_tick >= 0 {
-                     return None;
-                  }
-               }
-            }
-            TimingStep::Exhausted => break,
-            TimingStep::Invalid => return None,
-         }
-      }
-
-      if walker.has_unconsumed_ctts() || ticks_by_sample.len() != sample_count {
-         return None;
-      }
-      samples_by_time.sort_unstable_by_key(|sample_index| {
-         let index = usize::try_from(*sample_index).expect("u32 fits usize") - 1;
-         (ticks_by_sample[index], *sample_index)
-      });
-      Some(Self {
-         ticks_by_sample,
-         samples_by_time,
-      })
-   }
-
-   pub fn select(&self, target_tick: u64) -> Option<SampleSelection> {
-      let target = (i128::from(target_tick), u32::MAX);
-      let upper = self.samples_by_time.partition_point(|sample_index| {
-         let index = usize::try_from(*sample_index).expect("u32 fits usize") - 1;
-         (self.ticks_by_sample[index], *sample_index) <= target
-      });
-      let sample_index = if upper == 0 {
-         *self.samples_by_time.first()?
-      } else {
-         self.samples_by_time[upper - 1]
-      };
-      Some(SampleSelection {
-         sample_index,
-         presentation_tick: u64::try_from(self.tick(sample_index)?).ok()?,
-      })
-   }
-
-   pub fn tick(&self, sample_index: u32) -> Option<i128> {
-      let index = usize::try_from(sample_index).ok()?.checked_sub(1)?;
-      self.ticks_by_sample.get(index).copied()
-   }
-
-   pub fn ticks_for_range(&self, start_sample: u32, end_sample: u32) -> Option<Vec<(u32, i128)>> {
-      if start_sample > end_sample {
-         return None;
-      }
-      let start = usize::try_from(start_sample).ok()?.checked_sub(1)?;
-      let end = usize::try_from(end_sample).ok()?;
-      let ticks = self.ticks_by_sample.get(start..end)?;
-      let mut result = Vec::new();
-      result.try_reserve_exact(ticks.len()).ok()?;
-      for (offset, tick) in ticks.iter().copied().enumerate() {
-         let sample_index = start_sample.checked_add(u32::try_from(offset).ok()?)?;
-         result.push((sample_index, tick));
-      }
-      Some(result)
-   }
 }
 
 /// Reads the entry count of a full-box table (8-byte header of version/flags
@@ -291,260 +153,6 @@ pub fn parse_stss(stss: &[u8]) -> Option<Vec<u32>> {
       .then_some(samples)
 }
 
-pub fn parse_ctts(ctts: &[u8]) -> Option<Vec<CompositionOffset>> {
-   let version = *ctts.first()?;
-   if version > 1 {
-      return None;
-   }
-   let entry_count = table_entries(ctts, 8)?;
-
-   let mut offsets = Vec::new();
-   offsets.try_reserve(entry_count).ok()?;
-   for index in 0..entry_count {
-      let offset = 8 + index * 8;
-      let sample_count = read_u32_be(ctts, offset)?;
-      if sample_count == 0 {
-         return None;
-      }
-      let raw_offset = read_u32_be(ctts, offset + 4)?;
-      offsets.push(CompositionOffset {
-         sample_count,
-         sample_offset: if version == 0 {
-            i64::from(raw_offset)
-         } else {
-            i64::from(i32::from_be_bytes(raw_offset.to_be_bytes()))
-         },
-      });
-   }
-   Some(offsets)
-}
-
-/// One stts/ctts segment in decode order: `sample_count` samples starting at
-/// `first_sample` that share one sample delta and one composition offset.
-#[derive(Debug, Clone, Copy)]
-struct TimingSegment {
-   first_sample: u32,
-   decode_tick: u64,
-   sample_count: u32,
-   sample_delta: u64,
-   composition_offset: i64,
-}
-
-/// Result of advancing a [`TimingWalker`].
-enum TimingStep {
-   Segment(TimingSegment),
-   /// All stts entries were consumed.
-   Exhausted,
-   /// The timing tables are malformed (zero count/delta, missing ctts entry,
-   /// or tick overflow).
-   Invalid,
-}
-
-/// Walks the stts/ctts sample timing tables segment by segment in decode
-/// order, keeping the running sample index and decode tick.
-struct TimingWalker<'a> {
-   stts: &'a [u8],
-   composition_offsets: Option<&'a [CompositionOffset]>,
-   entry_count: usize,
-   sample_index: u32,
-   decode_tick: u64,
-   stts_index: usize,
-   stts_remaining: u32,
-   sample_delta: u64,
-   ctts_index: usize,
-   ctts_remaining: u32,
-   composition_offset: i64,
-}
-
-impl<'a> TimingWalker<'a> {
-   fn new(stts: &'a [u8], composition_offsets: Option<&'a [CompositionOffset]>) -> Option<Self> {
-      let entry_count = table_entries(stts, 8)?;
-      Some(Self {
-         stts,
-         composition_offsets,
-         entry_count,
-         sample_index: 1,
-         decode_tick: 0,
-         stts_index: 0,
-         stts_remaining: 0,
-         sample_delta: 0,
-         ctts_index: 0,
-         ctts_remaining: 0,
-         composition_offset: 0,
-      })
-   }
-
-   fn next_segment(&mut self) -> TimingStep {
-      if self.stts_remaining == 0 {
-         if self.stts_index == self.entry_count {
-            return TimingStep::Exhausted;
-         }
-         let offset = 8 + self.stts_index * 8;
-         let (Some(remaining), Some(delta)) = (
-            read_u32_be(self.stts, offset),
-            read_u32_be(self.stts, offset + 4),
-         ) else {
-            return TimingStep::Invalid;
-         };
-         if remaining == 0 || delta == 0 {
-            return TimingStep::Invalid;
-         }
-         self.stts_remaining = remaining;
-         self.sample_delta = u64::from(delta);
-         self.stts_index += 1;
-      }
-
-      if let Some(offsets) = self.composition_offsets {
-         if self.ctts_remaining == 0 {
-            let Some(entry) = offsets.get(self.ctts_index) else {
-               return TimingStep::Invalid;
-            };
-            self.ctts_remaining = entry.sample_count;
-            self.composition_offset = entry.sample_offset;
-            self.ctts_index += 1;
-         }
-      } else {
-         self.ctts_remaining = self.stts_remaining;
-         self.composition_offset = 0;
-      }
-
-      let segment_count = self.stts_remaining.min(self.ctts_remaining);
-      let segment = TimingSegment {
-         first_sample: self.sample_index,
-         decode_tick: self.decode_tick,
-         sample_count: segment_count,
-         sample_delta: self.sample_delta,
-         composition_offset: self.composition_offset,
-      };
-      let advance = u64::from(segment_count)
-         .checked_mul(self.sample_delta)
-         .and_then(|duration| self.decode_tick.checked_add(duration))
-         .and_then(|decode_tick| {
-            self
-               .sample_index
-               .checked_add(segment_count)
-               .map(|sample_index| (decode_tick, sample_index))
-         });
-      let Some((decode_tick, next_sample_index)) = advance else {
-         return TimingStep::Invalid;
-      };
-      self.decode_tick = decode_tick;
-      self.sample_index = next_sample_index;
-      self.stts_remaining -= segment_count;
-      self.ctts_remaining -= segment_count;
-      TimingStep::Segment(segment)
-   }
-
-   /// 1-based index of the next sample the walker will describe.
-   #[cfg(test)]
-   fn sample_index(&self) -> u32 {
-      self.sample_index
-   }
-
-   /// Whether any ctts-described samples remain after the stts table ended.
-   fn has_unconsumed_ctts(&self) -> bool {
-      self.ctts_remaining != 0
-         || self
-            .composition_offsets
-            .is_some_and(|offsets| self.ctts_index != offsets.len())
-   }
-}
-
-#[cfg(test)]
-pub fn select_sample_by_time(
-   stts: &[u8],
-   composition_offsets: Option<&[CompositionOffset]>,
-   presentation_offset: i64,
-   target_tick: u64,
-) -> Option<SampleSelection> {
-   let mut walker = TimingWalker::new(stts, composition_offsets)?;
-   let mut before: Option<SampleSelection> = None;
-   let mut after: Option<SampleSelection> = None;
-
-   loop {
-      match walker.next_segment() {
-         TimingStep::Segment(segment) => {
-            let first_presentation_tick = i128::from(segment.decode_tick)
-               .checked_add(i128::from(segment.composition_offset))?
-               .checked_sub(i128::from(presentation_offset))?;
-            consider_presentation_segment(
-               &mut before,
-               &mut after,
-               segment.first_sample,
-               first_presentation_tick,
-               segment.sample_count,
-               segment.sample_delta,
-               target_tick,
-            )?;
-         }
-         TimingStep::Exhausted => break,
-         TimingStep::Invalid => return None,
-      }
-   }
-
-   if walker.has_unconsumed_ctts() {
-      return None;
-   }
-
-   before.or(after)
-}
-
-/// Considers one presentation segment for sample selection.
-///
-/// The caller must guarantee `sample_delta` is non-zero (the stts/ctts walker
-/// rejects zero deltas while producing segments).
-#[cfg(test)]
-fn consider_presentation_segment(
-   before: &mut Option<SampleSelection>,
-   after: &mut Option<SampleSelection>,
-   first_sample: u32,
-   first_tick: i128,
-   sample_count: u32,
-   sample_delta: u64,
-   target_tick: u64,
-) -> Option<()> {
-   debug_assert!(sample_delta > 0);
-   let delta = i128::from(sample_delta);
-   let count = i128::from(sample_count);
-   let first_nonnegative = if first_tick < 0 {
-      (-first_tick).checked_add(delta.checked_sub(1)?)? / delta
-   } else {
-      0
-   };
-   if first_nonnegative >= count {
-      return Some(());
-   }
-
-   let first_valid_tick = first_tick.checked_add(first_nonnegative.checked_mul(delta)?)?;
-   let target = i128::from(target_tick);
-   if first_valid_tick <= target {
-      let position = ((target - first_tick) / delta).min(count - 1);
-      let presentation_tick = first_tick.checked_add(position.checked_mul(delta)?)?;
-      let candidate = SampleSelection {
-         sample_index: first_sample.checked_add(u32::try_from(position).ok()?)?,
-         presentation_tick: u64::try_from(presentation_tick).ok()?,
-      };
-      if before.is_none_or(|current| {
-         (candidate.presentation_tick, candidate.sample_index)
-            > (current.presentation_tick, current.sample_index)
-      }) {
-         *before = Some(candidate);
-      }
-   } else {
-      let candidate = SampleSelection {
-         sample_index: first_sample.checked_add(u32::try_from(first_nonnegative).ok()?)?,
-         presentation_tick: u64::try_from(first_valid_tick).ok()?,
-      };
-      if after.is_none_or(|current| {
-         (candidate.presentation_tick, candidate.sample_index)
-            < (current.presentation_tick, current.sample_index)
-      }) {
-         *after = Some(candidate);
-      }
-   }
-   Some(())
-}
-
 pub fn nearest_sync_sample(sample_index: u32, sync_samples: Option<&[u32]>) -> u32 {
    let Some(sync_samples) = sync_samples else {
       return sample_index;
@@ -576,31 +184,28 @@ pub fn next_sync_sample(
    }
 }
 
-#[cfg(test)]
-fn sample_file_offset(
-   sample_index: u32,
-   sizes: &SampleSizes,
-   stsc: &[StscEntry],
-   chunk_offsets: &[u64],
-) -> Option<u64> {
-   sample_location(sample_index, sizes, stsc, chunk_offsets).map(|location| location.file_offset)
-}
-
 pub fn sample_description_index(
    sample_index: u32,
    sizes: &SampleSizes,
    stsc: &[StscEntry],
    chunk_offsets: &[u64],
 ) -> Option<u32> {
-   sample_location(sample_index, sizes, stsc, chunk_offsets)
-      .map(|location| location.sample_description_index)
-}
+   if sample_index == 0
+      || sample_index > sizes.sample_count
+      || stsc.first()?.first_chunk != 1
+      || chunk_offsets.is_empty()
+   {
+      return None;
+   }
 
-#[derive(Debug, Clone, Copy)]
-struct SampleLocation {
-   #[cfg(test)]
-   file_offset: u64,
-   sample_description_index: u32,
+   let target = sample_index - 1;
+   for run in StscRuns::new(stsc, chunk_offsets.len())? {
+      let run_end = run.first_sample_index.checked_add(run.sample_count)?;
+      if target < run_end {
+         return Some(run.sample_description_index);
+      }
+   }
+   None
 }
 
 /// One stsc run: `chunk_count` consecutive chunks of `samples_per_chunk`
@@ -801,46 +406,6 @@ impl<'a> SampleLocator<'a> {
    }
 }
 
-fn sample_location(
-   sample_index: u32,
-   sizes: &SampleSizes,
-   stsc: &[StscEntry],
-   chunk_offsets: &[u64],
-) -> Option<SampleLocation> {
-   if sample_index == 0
-      || sample_index > sizes.sample_count
-      || stsc.first()?.first_chunk != 1
-      || chunk_offsets.is_empty()
-   {
-      return None;
-   }
-
-   let target = sample_index - 1;
-   for run in StscRuns::new(stsc, chunk_offsets.len())? {
-      let run_end = run.first_sample_index.checked_add(run.sample_count)?;
-      if target < run_end {
-         let within_run = target.checked_sub(run.first_sample_index)?;
-         let chunk_in_run = within_run / run.samples_per_chunk;
-         let within_chunk = within_run % run.samples_per_chunk;
-         let chunk_number = run.first_chunk.checked_add(chunk_in_run)?;
-         let first_sample_in_chunk = run
-            .first_sample_index
-            .checked_add(chunk_in_run.checked_mul(run.samples_per_chunk)?)?;
-         let prior_bytes = sum_sample_sizes(first_sample_in_chunk, within_chunk, sizes)?;
-         let _file_offset = chunk_offsets
-            .get(usize::try_from(chunk_number.checked_sub(1)?).ok()?)?
-            .checked_add(prior_bytes)?;
-         return Some(SampleLocation {
-            #[cfg(test)]
-            file_offset: _file_offset,
-            sample_description_index: run.sample_description_index,
-         });
-      }
-   }
-
-   None
-}
-
 pub fn sample_size(sample_index: u32, sizes: &SampleSizes) -> Option<u32> {
    if sample_index == 0 || sample_index > sizes.sample_count {
       return None;
@@ -853,34 +418,6 @@ pub fn sample_size(sample_index: u32, sizes: &SampleSizes) -> Option<u32> {
          .get(usize::try_from(sample_index - 1).ok()?)
          .copied()
    }
-}
-
-#[cfg(test)]
-pub fn validate_sample_range(
-   start_sample: u32,
-   end_sample: u32,
-   max_total_bytes: usize,
-) -> Result<usize> {
-   if start_sample == 0 || end_sample < start_sample {
-      return Err(MediaParserError::InvalidFormat(format!(
-         "invalid sample range: {start_sample}..={end_sample}"
-      )));
-   }
-
-   let capacity = usize::try_from(
-      end_sample
-         .checked_sub(start_sample)
-         .and_then(|count| count.checked_add(1))
-         .ok_or_else(|| MediaParserError::InvalidFormat("sample range too large".to_string()))?,
-   )
-   .map_err(|_| MediaParserError::InvalidFormat("sample range too large".to_string()))?;
-   let max_by_overhead = max_total_bytes / std::mem::size_of::<Vec<u8>>();
-   if capacity > MAX_SAMPLES_PER_RANGE || capacity > max_by_overhead {
-      return Err(MediaParserError::InvalidFormat(format!(
-         "sample range contains too many samples: {capacity}"
-      )));
-   }
-   Ok(capacity)
 }
 
 pub fn validate_sample_tables(
@@ -934,267 +471,8 @@ pub fn validate_sample_tables(
 }
 
 #[cfg(test)]
-pub fn presentation_ticks_for_range(
-   stts: &[u8],
-   composition_offsets: Option<&[CompositionOffset]>,
-   presentation_offset: i64,
-   start_sample: u32,
-   end_sample: u32,
-) -> Option<Vec<(u32, i128)>> {
-   validate_sample_range(start_sample, end_sample, usize::MAX).ok()?;
-   let range_len = usize::try_from(
-      end_sample
-         .checked_sub(start_sample)
-         .and_then(|count| count.checked_add(1))?,
-   )
-   .ok()?;
-   let mut walker = TimingWalker::new(stts, composition_offsets)?;
-
-   let mut result = Vec::new();
-   result.try_reserve(range_len).ok()?;
-   while walker.sample_index() <= end_sample {
-      let TimingStep::Segment(segment) = walker.next_segment() else {
-         return None;
-      };
-      let count = segment.sample_count;
-      let wanted_start = start_sample.saturating_sub(segment.first_sample).min(count);
-      let wanted_end = end_sample
-         .checked_add(1)?
-         .saturating_sub(segment.first_sample)
-         .min(count);
-      for position in wanted_start..wanted_end {
-         let sample_decode_tick = segment
-            .decode_tick
-            .checked_add(u64::from(position).checked_mul(segment.sample_delta)?)?;
-         let presentation_tick = i128::from(sample_decode_tick)
-            .checked_add(i128::from(segment.composition_offset))?
-            .checked_sub(i128::from(presentation_offset))?;
-         result.push((
-            segment.first_sample.checked_add(position)?,
-            presentation_tick,
-         ));
-      }
-   }
-
-   (result.len() == range_len).then_some(result)
-}
-
-pub fn duration_to_ticks(duration: Duration, timescale: u32) -> u64 {
-   let ticks = duration.as_nanos().saturating_mul(u128::from(timescale)) / 1_000_000_000;
-   u64::try_from(ticks).unwrap_or(u64::MAX)
-}
-
-fn stts_sample_count(stts: &[u8]) -> Option<u32> {
-   let entry_count = table_entries(stts, 8)?;
-   (0..entry_count).try_fold(0u32, |total, index| {
-      total.checked_add(read_u32_be(stts, 8 + index * 8)?)
-   })
-}
-
-/// Total duration in media ticks described by the stts table.
-pub fn stts_duration_ticks(stts: &[u8]) -> Option<u64> {
-   let entry_count = table_entries(stts, 8)?;
-   (0..entry_count).try_fold(0u64, |total, index| {
-      let count = u64::from(read_u32_be(stts, 8 + index * 8)?);
-      let delta = u64::from(read_u32_be(stts, 8 + index * 8 + 4)?);
-      total.checked_add(count.checked_mul(delta)?)
-   })
-}
-
-pub fn ticks_to_duration(ticks: u64, timescale: u32) -> Duration {
-   if timescale == 0 {
-      return Duration::ZERO;
-   }
-   let nanos = u128::from(ticks).saturating_mul(1_000_000_000) / u128::from(timescale);
-   Duration::from_nanos(u64::try_from(nanos).unwrap_or(u64::MAX))
-}
-
-fn sum_sample_sizes(start_sample: u32, count: u32, sizes: &SampleSizes) -> Option<u64> {
-   if count == 0 {
-      return Some(0);
-   }
-   if sizes.fixed_size != 0 {
-      return u64::from(sizes.fixed_size).checked_mul(u64::from(count));
-   }
-
-   let start = usize::try_from(start_sample).ok()?;
-   let end = start.checked_add(usize::try_from(count).ok()?)?;
-   sizes
-      .sizes
-      .get(start..end)?
-      .iter()
-      .try_fold(0u64, |total, size| total.checked_add(u64::from(*size)))
-}
-
-#[cfg(test)]
 mod tests {
    use super::*;
-
-   #[test]
-   fn selects_sample_without_expanding_stts() {
-      let mut stts = vec![0; 8];
-      stts[4..8].copy_from_slice(&1u32.to_be_bytes());
-      stts.extend_from_slice(&4u32.to_be_bytes());
-      stts.extend_from_slice(&1_000u32.to_be_bytes());
-
-      let selection = select_sample_by_time(&stts, None, 0, 2_500).unwrap();
-
-      assert_eq!(selection.sample_index, 3);
-      assert_eq!(selection.presentation_tick, 2_000);
-   }
-
-   #[test]
-   fn selects_sample_by_ctts_presentation_time() {
-      let mut stts = vec![0; 8];
-      stts[4..8].copy_from_slice(&1u32.to_be_bytes());
-      stts.extend_from_slice(&3u32.to_be_bytes());
-      stts.extend_from_slice(&1_000u32.to_be_bytes());
-
-      let mut ctts = vec![0; 8];
-      ctts[4..8].copy_from_slice(&3u32.to_be_bytes());
-      for offset in [2_000u32, 3_000, 1_000] {
-         ctts.extend_from_slice(&1u32.to_be_bytes());
-         ctts.extend_from_slice(&offset.to_be_bytes());
-      }
-      let composition_offsets = parse_ctts(&ctts).unwrap();
-
-      let selection =
-         select_sample_by_time(&stts, Some(&composition_offsets), 2_000, 1_000).unwrap();
-
-      assert_eq!(selection.sample_index, 3);
-      assert_eq!(selection.presentation_tick, 1_000);
-   }
-
-   #[test]
-   fn presentation_timeline_matches_full_walker_for_reordered_and_equal_ticks() {
-      let mut stts = vec![0; 8];
-      stts[4..8].copy_from_slice(&1u32.to_be_bytes());
-      stts.extend_from_slice(&5u32.to_be_bytes());
-      stts.extend_from_slice(&1_000u32.to_be_bytes());
-      let composition_offsets = vec![
-         CompositionOffset {
-            sample_count: 1,
-            sample_offset: 2_000,
-         },
-         CompositionOffset {
-            sample_count: 1,
-            sample_offset: 0,
-         },
-         CompositionOffset {
-            sample_count: 1,
-            sample_offset: -2_000,
-         },
-         CompositionOffset {
-            sample_count: 1,
-            sample_offset: 0,
-         },
-         CompositionOffset {
-            sample_count: 1,
-            sample_offset: -1_000,
-         },
-      ];
-      let timeline =
-         PresentationTimeline::new(&stts, Some(&composition_offsets), 1_000, 5).unwrap();
-
-      for target in [0, 500, 1_000, 1_500, 2_000, 2_500, 10_000] {
-         assert_eq!(
-            timeline.select(target),
-            select_sample_by_time(&stts, Some(&composition_offsets), 1_000, target),
-            "target {target}"
-         );
-      }
-
-      let shifted = PresentationTimeline::new(&stts, None, -1_000, 5).unwrap();
-      assert_eq!(
-         shifted.select(0),
-         select_sample_by_time(&stts, None, -1_000, 0)
-      );
-   }
-
-   #[test]
-   fn presentation_timeline_preserves_signed_ticks_in_decode_order_ranges() {
-      let mut stts = vec![0; 8];
-      stts[4..8].copy_from_slice(&1u32.to_be_bytes());
-      stts.extend_from_slice(&3u32.to_be_bytes());
-      stts.extend_from_slice(&1_000u32.to_be_bytes());
-      let timeline = PresentationTimeline::new(&stts, None, 1_500, 3).unwrap();
-
-      assert_eq!(
-         timeline.ticks_for_range(1, 3),
-         Some(vec![(1, -1_500), (2, -500), (3, 500)])
-      );
-      assert_eq!(
-         timeline.ticks_for_range(1, 3),
-         presentation_ticks_for_range(&stts, None, 1_500, 1, 3)
-      );
-      assert_eq!(timeline.tick(2), Some(-500));
-      assert_eq!(timeline.ticks_for_range(0, 1), None);
-      assert_eq!(timeline.ticks_for_range(2, 1), None);
-      assert_eq!(timeline.ticks_for_range(1, 4), None);
-   }
-
-   #[test]
-   fn presentation_timeline_rejects_malformed_timing_tables() {
-      fn stts(count: u32, delta: u32) -> Vec<u8> {
-         let mut bytes = vec![0; 8];
-         bytes[4..8].copy_from_slice(&1u32.to_be_bytes());
-         bytes.extend_from_slice(&count.to_be_bytes());
-         bytes.extend_from_slice(&delta.to_be_bytes());
-         bytes
-      }
-
-      assert!(PresentationTimeline::new(&stts(0, 1), None, 0, 1).is_none());
-      assert!(PresentationTimeline::new(&[], None, 0, 0).is_none());
-      assert!(PresentationTimeline::new(&[], None, 0, 1).is_none());
-      assert!(PresentationTimeline::new(&stts(1, 0), None, 0, 1).is_none());
-      assert!(
-         PresentationTimeline::new(
-            &stts(2, 1),
-            Some(&[CompositionOffset {
-               sample_count: 1,
-               sample_offset: 0,
-            }]),
-            0,
-            2,
-         )
-         .is_none()
-      );
-      assert!(
-         PresentationTimeline::new(
-            &stts(1, 1),
-            Some(&[CompositionOffset {
-               sample_count: 2,
-               sample_offset: 0,
-            }]),
-            0,
-            1,
-         )
-         .is_none()
-      );
-      assert!(PresentationTimeline::new(&stts(1, 1), None, 0, 2).is_none());
-      assert!(PresentationTimeline::new(&stts(2, u32::MAX), None, 0, 2).is_some());
-      assert!(PresentationTimeline::new(&stts(u32::MAX, u32::MAX), None, 0, u32::MAX).is_none());
-   }
-
-   #[test]
-   fn presentation_timeline_pins_the_storage_budget_boundary() {
-      let bytes_per_sample = std::mem::size_of::<i128>() + std::mem::size_of::<u32>();
-      let max_samples = MAX_PRESENTATION_TIMELINE_BYTES / bytes_per_sample;
-
-      assert!(!presentation_timeline_sample_count_fits(0));
-      assert!(presentation_timeline_sample_count_fits(max_samples));
-      assert!(!presentation_timeline_sample_count_fits(max_samples + 1));
-   }
-
-   #[test]
-   fn rejects_zero_sample_delta_while_selecting() {
-      let mut stts = vec![0; 8];
-      stts[4..8].copy_from_slice(&1u32.to_be_bytes());
-      stts.extend_from_slice(&2u32.to_be_bytes());
-      stts.extend_from_slice(&0u32.to_be_bytes());
-
-      assert!(select_sample_by_time(&stts, None, 0, 1_000).is_none());
-   }
 
    fn two_run_tables() -> (SampleSizes, Vec<StscEntry>, Vec<u64>) {
       // chunks 1-2 use description 1 (2 samples each), chunks 3-4 use
@@ -1258,14 +536,15 @@ mod tests {
    }
 
    #[test]
-   fn sample_locator_matches_sample_file_offset_in_order() {
+   fn sample_locator_matches_expected_offsets_in_order() {
       let (sizes, stsc, chunk_offsets) = two_run_tables();
       let mut locator = SampleLocator::new(&sizes, &stsc, &chunk_offsets).unwrap();
 
-      for sample_index in 1..=6 {
+      for (sample_index, expected_offset) in [100, 110, 200, 230, 300, 400].into_iter().enumerate()
+      {
          assert_eq!(
-            locator.file_offset(sample_index),
-            sample_file_offset(sample_index, &sizes, &stsc, &chunk_offsets),
+            locator.file_offset(u32::try_from(sample_index + 1).unwrap()),
+            Some(expected_offset),
             "sample {sample_index}"
          );
       }
@@ -1281,43 +560,6 @@ mod tests {
    }
 
    #[test]
-   fn sums_stts_duration_with_checked_arithmetic() {
-      let mut stts = vec![0; 8];
-      stts[4..8].copy_from_slice(&2u32.to_be_bytes());
-      stts.extend_from_slice(&3u32.to_be_bytes());
-      stts.extend_from_slice(&1_000u32.to_be_bytes());
-      stts.extend_from_slice(&2u32.to_be_bytes());
-      stts.extend_from_slice(&500u32.to_be_bytes());
-
-      assert_eq!(stts_duration_ticks(&stts), Some(4_000));
-
-      let mut large = vec![0; 8];
-      large[4..8].copy_from_slice(&1u32.to_be_bytes());
-      large.extend_from_slice(&u32::MAX.to_be_bytes());
-      large.extend_from_slice(&u32::MAX.to_be_bytes());
-
-      assert_eq!(
-         stts_duration_ticks(&large),
-         Some(u64::from(u32::MAX) * u64::from(u32::MAX))
-      );
-   }
-
-   #[test]
-   fn parses_signed_ctts_v1_offsets() {
-      let mut ctts = vec![1, 0, 0, 0, 0, 0, 0, 1];
-      ctts.extend_from_slice(&2u32.to_be_bytes());
-      ctts.extend_from_slice(&(-500i32).to_be_bytes());
-
-      assert_eq!(
-         parse_ctts(&ctts).unwrap(),
-         vec![CompositionOffset {
-            sample_count: 2,
-            sample_offset: -500,
-         }]
-      );
-   }
-
-   #[test]
    fn preserves_stsc_sample_description_index() {
       let mut stsc = vec![0; 8];
       stsc[4..8].copy_from_slice(&1u32.to_be_bytes());
@@ -1328,11 +570,6 @@ mod tests {
       let entries = parse_stsc(&stsc).unwrap();
 
       assert_eq!(entries[0].sample_description_index, 3);
-   }
-
-   #[test]
-   fn rejects_hostile_sample_range_before_allocating() {
-      assert!(validate_sample_range(1, u32::MAX, 64 * 1024 * 1024).is_err());
    }
 
    #[test]
@@ -1350,12 +587,12 @@ mod tests {
       let chunk_offsets = vec![0];
 
       assert_eq!(
-         sample_file_offset(1_000, &sizes, &stsc, &chunk_offsets),
-         Some(3_996)
+         sample_description_index(1_000, &sizes, &stsc, &chunk_offsets),
+         Some(1)
       );
       assert_eq!(
-         sample_file_offset(1_000_000_000, &sizes, &stsc, &chunk_offsets),
-         Some(3_999_999_996)
+         sample_description_index(1_000_000_000, &sizes, &stsc, &chunk_offsets),
+         Some(1)
       );
    }
